I'm going to have to figure out and back out some of this stuff later, but honestly if I tried to take the time to fully work out all the boring boilerplate bullshit of Android app buildsystems I'm sure I'd just burn out quickly on this.
So lets just ignore all that for now; I've at least got it so it'll build on my setup of Android Studio and push successfully and run on my watch over wireless adb. Now I can start to try and work out why the watchface isn't taking up the entire display on newer watches; is it the display density, or is it the target Android Wear version? How does one actually set or change the target version? Oh god it's just Android buildsystem bullshit all the way down, isn't it . . .